Add files via upload
Priority  Queue:  Suppose  you  are  using  the  priority  queue  data  structure to maintain plane-landing system. As planes arrive near the vicinity of the airport, each plane sends a landing request with the duration(in minutes)that the plane can wait before landing.  For example, plane may send the request with value 30. That means that the plane can wait 30 minutes before landing.  Once a landing request is received, the request will be added to the  plane  landing  priority  queue.  Each  time  when  the  land  method  is  called,  the  landing request with lowest wait time will be removed from the priority queue. Develop a C or C++ program that implements this plane landing system with following user interface

Welcome to the Plane Landing System

1. Make a landing request
2. Land a Plane
3. List all the landing requests
4. Exit
Select your option: